CLEANUP
turns out code is a lot cleaner when you dont need to do iloc random crimes
