+++ chipdata.tsv
@@ -12,7 +12,7 @@ ICEPick	2	Close	Can instead be used to control inanimate software (e.g. control 
 IceSkates	1	Close	A frosty kick upshifts target’s next Body roll! Also, you can now freely glide over ice.	Blade			A frosty kick upshifts target’s next Body roll! Also, you can now freely glide over ice.	Core		
 Katana	2x2	Close	Double-slice! If you win a parry with this, the target cannot use chips until they roll 3 times.	Blade	Awkward	 [Awkward]	Double-slice! If you win a parry with this, the target cannot use chips until they roll 3 times. [Awkward]	Core		
 KnightSword	2	Close	Deals double damage if this chip wins a parry.	Blade			Deals double damage if this chip wins a parry.	Core		
-LaserRazor	3	Close	A glowing razor made of raw energy. It can gradually burn through any material.	Blade	Dangerous, Slow	 [Dangerous, Slow]	A glowing razor made of raw energy. It can gradually burn through any material. [Dangerous, Slow]	Core	S-License	
+LaserRazor	3	Close	A glowing razor made of raw energy. It can gradually burn through any material.	Blade	Dangerous, Slow	 [Dangerous, Slow]	A glowing razor made of raw energy. It can gradually burn through any material. [Dangerous, Slow]	Core	S-License	LazerRazor, LazerRazer
 LightningFoil	1	Close	An electric disarming swipe removes the first alphabetic BattleChip in the target’s Folder.	Blade	Accurate	 [Accurate]	An electric disarming swipe removes the first alphabetic BattleChip in the target’s Folder. [Accurate]	Core		
 LightWave	2	Near	Fires a glowing shockwave from a blade that bypasses Guards.	Blade			Fires a glowing shockwave from a blade that bypasses Guards.	Core		
 MirrorSword	2	Close	A shining silver sword. Can instead be used to parry a projectile attack.	Blade			A shining silver sword. Can instead be used to parry a projectile attack.	Core	A-License
