+++ maincommon.py
@@ -103,3 +103,18 @@ async def find_value_in_table(context, df, search_col, search_arg, suppress_notf
                                      sendcontent="Found more than one match for %s! You should probably let the devs know..." % search_arg)
         return None
     return search_results.iloc[0]
+
+
+def roll_row_from_table(roll_df, df_filters={}):
+    bool_filt = None
+    if df_filters:
+        for k, v in df_filters.items():
+            if bool_filt is None:
+                bool_filt = (roll_df[k] == v)
+            else:
+                bool_filt = bool_filt & (roll_df[k] == v)
+        sub_df = roll_df[bool_filt]
+    else:
+        sub_df = roll_df
+    row_num = random.randint(1, sub_df.shape[0]) - 1
+    return sub_df.iloc[row_num]
