+++ mysterydata.tsv
@@ -441,3 +441,35 @@ Sapphire	Misc Table	The Mystery Data is a hidden Prog. It has a risky request; c
 Sapphire	Misc Table	The Navi gains an egg. After time and care, it hatches into a friendly Virus. It’s a big responsibility.
 Sapphire	Misc Table	The Navi can ask for any item. They’ll learn the (non-monetary) price. When paid, they get the item.
 Sapphire	Misc Table	The Mystery Data shimmers, then duplicates. The two MDs will only open when given as a gift.
+		
+Sunny	Info	BattleChip
+Sunny	Info	BattleChip
+Sunny	Info	BattleChip
+Sunny	Info	Bonus Roll: Roll 2 more times and get both!
+Sunny	BattleChip	BlastRocket
+Sunny	BattleChip	BlinkSlash
+Sunny	BattleChip	BullRush
+Sunny	BattleChip	Camouflage
+Sunny	BattleChip	ColorBall
+Sunny	BattleChip	CutContent
+Sunny	BattleChip	DimensionalSlice
+Sunny	BattleChip	Discord
+Sunny	BattleChip	DoubleJump
+Sunny	BattleChip	LightEdge
+Sunny	BattleChip	Lookout
+Sunny	BattleChip	Magnifryer
+Sunny	BattleChip	Masquerade
+Sunny	BattleChip	MoonBlade
+Sunny	BattleChip	Nailgun
+Sunny	BattleChip	ProtoStrike
+Sunny	BattleChip	RainBow
+Sunny	BattleChip	RemotePlug
+Sunny	BattleChip	RollArrow
+Sunny	BattleChip	SeismicBlast
+Sunny	BattleChip	ShadowFreeze
+Sunny	BattleChip	ShieldBash
+Sunny	BattleChip	StarArrow
+Sunny	BattleChip	StatiCannon
+Sunny	BattleChip	Underhand
+Sunny	BattleChip	Uplink
+Sunny	BattleChip	WirelessMouse
